backend/requirements.txt
alembic==1.17.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
backoff==2.2.1
backports.asyncio.runner==1.2.0
bcrypt==5.0.0
billiard==4.2.2
black==25.9.0
boolean.py==5.0
build==1.3.0
CacheControl==0.14.3
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
chromadb==1.2.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coloredlogs==15.0.1
coverage==7.11.0
cryptography==46.0.3
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
durationpy==0.10
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.119.1
filelock==3.20.0
flake8==7.3.0
flatbuffers==25.9.23
fsspec==2025.9.0
google-auth==2.41.1
googleapis-common-protos==1.71.0
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
kubernetes==34.1.0
license-expression==30.4.4
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
mypy==1.18.2
mypy_extensions==1.1.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.1
openai==2.6.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
overrides==7.7.0
packageurl-python==0.17.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pillow==12.0.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.5.0
pluggy==1.6.0
posthog==5.4.0
pre_commit==4.3.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.0
py-serializable==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.5
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
starlette==0.48.0
structlog==25.4.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
virtualenv==20.35.3
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
zipp==3.23.0
google-generativeai
PyGithub
figma-api
